assignment01 PartA unit test implemented

+++ assignment01/PartA/hadoop_exe.sh
@@ -1,15 +1,13 @@
-javac -classpath $HADOOP_HOME/hadoop-core-0.20.203.0.jar -d classes src/edu/duke/cs/AverageTemperature.java 
+if [ $# -ne 1 ] ; then
+	echo "command input_filename"
+	exit
+fi
 
-jar -cvf jar/AverageTemperature.jar -C classes/ .
-
-hadoop fs -rmr ~/PartA_input
-hadoop fs -mkdir ~/PartA_input
-hadoop dfs -copyFromLocal data/file01 ~/PartA_input/
-hadoop dfs -copyFromLocal data/file02 ~/PartA_input/
+jar -cvf jar/AverageTemperature.jar -C bin/ .
 
 hadoop fs -rmr ~/PartA_output
 
-hadoop jar jar/AverageTemperature.jar edu/duke/cs/AverageTemperature  ~/PartA_input/file* ~/PartA_output
+hadoop jar jar/AverageTemperature.jar edu/duke/cs/AverageTemperatureDriver  ~/PartA_input/$1 ~/PartA_output
   
 hadoop fs -lsr ~/PartA_output
